requirements.txt
git+https://github.com/garulf/flox#egg=flox
pygithub==1.55
requests==2.25.1